first version of fixed station settings on subfragment
