+++ app/src/main/java/hpsaturn/pollutionreporter/view/SettingsDeviceTools.java
@@ -5,10 +5,13 @@ import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 
+import hpsaturn.pollutionreporter.R;
+import hpsaturn.pollutionreporter.models.ResponseConfig;
+
 /**
  * Created by Antonio Vanegas @hpsaturn on 3/31/21.
  */
-public class SettingsMobileStation extends SettingsBaseFragment{
+public class SettingsDeviceTools extends SettingsBaseFragment{
 
 
     @Override
@@ -18,6 +21,19 @@ public class SettingsMobileStation extends SettingsBaseFragment{
 
     @Override
     public void onCreatePreferencesFix(@Nullable Bundle savedInstanceState, String rootKey) {
+        setPreferencesFromResource(R.xml.settings_device_tools, rootKey);
 
     }
+
+    @Override
+    protected void refreshUI() {
+
+    }
+
+    @Override
+    protected void onConfigRead(ResponseConfig config) {
+
+    }
+
+
 }
