+++ app/src/main/res/xml/settings_device_tools.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <PreferenceCategory app:title="Device Status">
+        <SwitchPreference
+            app:icon="@android:drawable/ic_lock_power_off"
+            app:key="@string/key_device_status"
+            app:summary="Disconnected"
+            app:defaultValue="false"
+            app:title="Device Status:" />
+    </PreferenceCategory>
+
+
+    <PreferenceCategory app:title="Some Device Tools"
+        android:dependency="@string/key_device_status">
+
+        <SwitchPreference
+            app:key="@string/key_setting_enable_location"
+            app:title="Save Current Location" />
+
+        <SwitchPreference
+            app:defaultValue="false"
+            app:key="@string/key_setting_enable_reboot"
+            app:summary="@string/summary_reboot"
+            app:title="Reboot Device" />
+
+        <SwitchPreference
+            app:defaultValue="false"
+            app:key="@string/key_setting_enable_clear"
+            app:summary="@string/summary_clear"
+            app:title="Factory Reset" />
+
+    </PreferenceCategory>
+
+</PreferenceScreen>
