+++ app/src/main/res/xml/settings_fixed_station.xml
@@ -2,9 +2,39 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <PreferenceCategory app:title="Device Status">
+        <SwitchPreference
+            app:defaultValue="false"
+            app:icon="@android:drawable/ic_lock_power_off"
+            app:key="@string/key_device_status"
+            app:summary="Disconnected"
+            app:title="Device Status:" />
+    </PreferenceCategory>
+
     <PreferenceCategory
-        app:title="Fixed Station Settings"
-        android:dependency="@string/key_device_status">
+        android:dependency="@string/key_device_status"
+        app:title="Mobile Station Settings">
+
+        <EditTextPreference
+            android:inputType="number"
+            app:defaultValue="5"
+            app:icon="@android:drawable/ic_menu_save"
+            app:key="@string/key_setting_stime"
+            app:summary="@string/summary_stime"
+            app:title="Sample Time Interval" />
+
+        <ListPreference
+            android:defaultValue="0"
+            android:entries="@array/pref_dtype_entries"
+            android:entryValues="@array/pref_dtype_values"
+            android:key="@string/key_setting_dtype"
+            android:title="Sensor type"
+            app:useSimpleSummaryProvider="true" />
+    </PreferenceCategory>
+
+    <PreferenceCategory
+        android:dependency="@string/key_device_status"
+        app:title="Fixed Station Settings">
 
         <EditTextPreference
             android:inputType="text"
@@ -14,15 +44,15 @@
             app:title="Station Name" />
 
         <EditTextPreference
-            app:dependency="@string/key_setting_dname"
             android:inputType="text"
+            app:dependency="@string/key_setting_dname"
             app:key="@string/key_setting_ssid"
             app:summary=""
             app:title="WiFi Name" />
 
         <EditTextPreference
-            app:dependency="@string/key_setting_dname"
             android:inputType="textPassword"
+            app:dependency="@string/key_setting_dname"
             app:key="@string/key_setting_pass"
             app:summary="WiFi password"
             app:title="Password" />
@@ -65,8 +95,8 @@
     </PreferenceCategory>
 
     <PreferenceCategory
-        app:isPreferenceVisible="false"
         app:dependency="@string/key_setting_enable_wifi"
+        app:isPreferenceVisible="false"
         app:title="Deprecated CanAirIO API">
 
         <EditTextPreference
@@ -83,16 +113,16 @@
             app:title="Password" />
 
         <EditTextPreference
+            android:defaultValue="@string/value_apisrv_default"
             android:inputType="text"
             app:key="@string/key_setting_apisrv"
-            android:defaultValue="@string/value_apisrv_default"
             app:summary=""
             app:title="Hostname" />
 
         <EditTextPreference
+            android:defaultValue="@string/value_apiuri_default"
             android:inputType="text"
             app:key="@string/key_setting_apiuri"
-            android:defaultValue="@string/value_apiuri_default"
             app:summary=""
             app:title="API target" />
 
@@ -102,8 +132,9 @@
 
     </PreferenceCategory>
 
-    <PreferenceCategory app:title="Optional Device Settings"
-        android:dependency="@string/key_device_status">
+    <PreferenceCategory
+        android:dependency="@string/key_device_status"
+        app:title="Optional Device Settings">
 
         <SwitchPreference
             app:key="@string/key_setting_enable_location"
@@ -124,8 +155,9 @@
 
     </PreferenceCategory>
 
-    <PreferenceCategory app:title=""
-        app:isPreferenceVisible="false">
+    <PreferenceCategory
+        app:isPreferenceVisible="false"
+        app:title="">
         <Preference
             app:icon="@android:drawable/ic_menu_send"
             app:key="@string/key_send_feedback"
@@ -140,7 +172,6 @@
             app:summary=""
             app:title="Device info:" />
     </PreferenceCategory>
-
 
 
 </PreferenceScreen>
