+++ gradle.properties
@@ -11,7 +11,7 @@ org.gradle.jvmargs=-Xmx1536m
 mCompileSdkVersion=29
 mMinSdkVersion=19
 mTargetSdkVersion=29
-mVersionCode=543
-mVersionName=0.4.0
+mVersionCode=544
+mVersionName=0.4.1
 android.useAndroidX=true
 android.enableJetifier=true
